+++ src/main/java/br/com/cast/avaliacao/service/CourseService.java
@@ -5,6 +5,8 @@ import java.util.Optional;
 
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.MessageSource;
+import org.springframework.context.i18n.LocaleContextHolder;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -18,7 +20,10 @@ public class CourseService {
 	@Autowired
 	private Courses courses;
 	
-	public Optional<Course> findById(Long id){
+	@Autowired
+	private MessageSource messageSource;
+
+	public Optional<Course> findById(Long id) {
 		return this.courses.findById(id);
 	}
 
@@ -29,13 +34,15 @@ public class CourseService {
 	public Page<Course> findPage(String description, Pageable pageable) {
 		return this.courses.findPage(description, pageable);
 	}
-	
+
 	public void delete(Long id) {
 		Optional<Course> optionalCourse = this.findById(id);
 		if (optionalCourse.isPresent()) {
-			this.courses.deleteById(id);			
+			this.courses.deleteById(id);
 		} else {
-			throw new BusinessException("Curso Inválido para exclusão.");
+			String userMessage = this.messageSource.getMessage("message.course-invalid-id", null,
+					LocaleContextHolder.getLocale());
+			throw new BusinessException(userMessage);
 		}
 	}
 
@@ -45,8 +52,10 @@ public class CourseService {
 		} else {
 			if (checkCourseDates(course)) {
 				return this.courses.save(course);
-			} else {				
-				throw new BusinessException("Existe(m) curso(s) planejado(s) dentro do mesmo período.");		
+			} else {
+				String userMessage = messageSource.getMessage("message.course-already-planned", null,
+						LocaleContextHolder.getLocale());
+				throw new BusinessException(userMessage);
 			}
 		}
 	}
@@ -58,22 +67,37 @@ public class CourseService {
 			if (checkCourseDates(course)) {
 				Optional<Course> optionalCourse = this.findById(id);
 				if (!optionalCourse.isPresent()) {
-					throw new BusinessException("Curso Inválido para exclusão.");
+					String userMessage = messageSource.getMessage("message.course-invalid-id", null,
+							LocaleContextHolder.getLocale());
+					throw new BusinessException(userMessage);
 				}
 				Course persistedCourse = optionalCourse.get();
-				BeanUtils.copyProperties(course, persistedCourse, "codigo");	
-				return this.courses.save(persistedCourse);
-			} else {				
-				throw new BusinessException("Existe(m) curso(s) planejado(s) dentro do mesmo período.");		
+				BeanUtils.copyProperties(course, persistedCourse, "codigo");
+				this.courses.save(persistedCourse);
+				return this.findById(persistedCourse.getId()).get();
+			} else {
+				String userMessage = messageSource.getMessage("message.course-already-planned", null,
+						LocaleContextHolder.getLocale());
+				throw new BusinessException(userMessage);
 			}
 		}
 	}
-	
-	private boolean checkCourseDates(Course course) {
+
+	public boolean checkCourseDates(Course course) {
 		if (course.getStartDate().isAfter(course.getFinishDate())) {
-			throw new BusinessException("A data de início do curoso é maior que a data de finalização.");
+			String userMessage = messageSource.getMessage("message.course-invalid-period", null,
+					LocaleContextHolder.getLocale());
+			throw new BusinessException(userMessage);
 		}
 		return this.courses.checkDates(course);
+	}
+
+	public MessageSource getMessageSource() {
+		return messageSource;
+	}
+
+	public void setMessageSource(MessageSource messageSource) {
+		this.messageSource = messageSource;
 	}
 
 }